Update ip_function_spaces.md
I believe the proof that the minimal norm solution is unique is incorrect. The main reason is that we have u = v + z where z must lie in the nullspace of K due to the normal equations. However, we know that z = u - v is in the orthogonal complements of this nullspace (by definition of u and v). So z = u - v = 0 (as it lies both in the nullspace and its orthogonal complement). For me it seems weird to immediately conclude from the normal equations that u - v = z must lie in the nullspace of K. I do not think this is true in general. 

In my change, I proposed a different proof of the statement.