Update ip_function_spaces.md
I added a statement that shows that solutions to the normal equation solve the minimal residual problem. Before we only had that the minimal residual solutions should satisfy the normal equations, but not the other way around. Hence, by my addition we have:

- An element u in U solves the minimal residual problem IF AND ONLY IF it is a solution to the normal equations.
- The minimal residual problem has a solution IF AND ONLY IF the right hand side f is in R(K) + R(K)^perp
- An element u in U is the minimum norm solution to the minimal residual problem IF AND ONLY IF it is the minimal norm solution to the normal equations.

Before my addition, we did not have the first point and hence also not the third bullet point. The third bullet point is important later on as then we show that the Moore Penrose Pseudo inverse is the unique minimum norm solution to the minimal residual problem. Before, we only showed that the unique minimal norm solution to the normal equations is given by the Moore Penrose Pseudo Inverse. Via my addition (see the third bullet point), this means that we also have the unique minimal norm solution to the minimal residual problem